+++ packages/expo-router/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Add partial support for `generateStaticParams` in React Server Components router. ([#34093](https://github.com/expo/expo/pull/34093) by [@EvanBacon](https://github.com/EvanBacon))
 - Add server error handling to error boundary. ([#33971](https://github.com/expo/expo/pull/33971) by [@EvanBacon](https://github.com/EvanBacon))
+- Add `router.preload(<href>)`, `<Preload href={}>`, and `<Link preload href={}>`
 
 ### 🐛 Bug fixes